Large reworking of linters/stylers. CreateSQL styler is much better at catching mistakes.

+++ AppRefiner/Stylers/PropertyAsVariable.cs
@@ -80,8 +80,8 @@ namespace AppRefiner.Stylers
                 {
                     Highlights.Add(new CodeHighlight
                     {
-                        Start = (uint)userVariable.Symbol.StartIndex,
-                        Length = (uint)userVariable.Symbol.Text.Length,
+                        Start = userVariable.Symbol.StartIndex,
+                        Length = userVariable.Symbol.Text.Length,
                         Color = HighlightColor.Salmon
                     });
                 }
